SMIL fractional repeatCount and capped accumulate iterations

+++ ksvg/src/main/kotlin/hu/oandras/ksvg/dom/animation/AnimateDashArray.kt
@@ -39,7 +39,7 @@ internal class AnimateDashArray(
     val by: FloatArray?,
     durMs: Long,
     beginMs: Long,
-    repeatCount: Int,
+    repeatCount: Float,
     repeatDurMs: Long,
     endMs: Long,
     fillFreeze: Boolean,
