+++ ksvg/src/main/kotlin/hu/oandras/ksvg/dom/animation/AnimateRotate.kt
@@ -26,7 +26,7 @@ internal class AnimateRotate(
     by: FloatList?,
     durMs: Long,
     beginMs: Long,
-    repeatCount: Int,
+    repeatCount: Float,
     repeatDurMs: Long,
     endMs: Long,
     fillFreeze: Boolean,
